Updates to authentication
The server login endpoint now returns a user object on success, in addition to setting
the JWT token cookies. This user object is stored in the client localstorage and is
used to determine if the user is logged in or not in the front end.

On subsequent requests, the server checks whether the access token is expired, and if so
response with a 403 error status. When the client recieves a 403 error, it will query the refresh-token endpoint
in order to get a fresh access token and store it as an httpOnly cookie again. The client app automatically
redirects to the route that the user went to with the expired token, creating a seamless token refresh flow.

There are also updates to the error handling to correct some async errors which were not being handled correctly.